Final Push.
Assalamualikum, I have done completing the assignment. There are two type of users in the db, There is a table named as 'admins', it is used to hold the users data. And email is sent to both of the users as per the creation and closure of the ticket. I have use SMTP with tls configuration here. Thank you.